--- typescript/src/files.ts
@@ -1,18 +1,0 @@
-import { promises as fs } from "fs";
-
-// Define a type for JSON data
-export type JsonData = Record<string, any>;
-
-export async function readFileReturnJson(filePath: string): Promise<JsonData[]> {
-
-    console.log(`Reading JSON file from ${filePath}`);
-
-    const fileAsString = await fs.readFile(filePath, "utf-8");
-    return JSON.parse(fileAsString);
-}
-export async function writeFileJson(filePath: string, jsonData: JsonData): Promise<void> {
-    const jsonString = JSON.stringify(jsonData, null, 2);
-    await fs.writeFile(filePath, jsonString, "utf-8");
-
-    console.log(`Wrote JSON file to ${filePath}`);
-}
